+++ lib/widgets/category_item_widget.dart
@@ -3,12 +3,10 @@ import 'dart:io';
 import 'package:drama_app/models/item.dart';
 import 'package:drama_app/providers/auth_provider.dart';
 import 'package:drama_app/providers/items_provider.dart';
-import 'package:drama_app/screens/auth_screen.dart';
 import 'package:drama_app/screens/items_screen.dart';
-import 'package:drama_app/screens/sign_btn_screen.dart';
+
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../screens/comment_screen.dart';
 
 import 'package:http/http.dart' as http;
 
@@ -21,13 +19,21 @@ class ItemWidget extends StatelessWidget {
   final Item wholeItem;
   final String trailerVideoUrl;
 
-  ItemWidget({required this.wholeItem, required this.id, required this.title, required this.imageUrls, required this.category, required this.genres, required this.trailerVideoUrl});
+  ItemWidget(
+      {required this.wholeItem,
+      required this.id,
+      required this.title,
+      required this.imageUrls,
+      required this.category,
+      required this.genres,
+      required this.trailerVideoUrl});
 
   void selectItemDetails(BuildContext ctx, String token) {
     Navigator.of(ctx).push(
       MaterialPageRoute(
         builder: (_) {
-          return ItemDetailsScreen(id, title, category, imageUrls, trailerVideoUrl, wholeItem, token);
+          return ItemDetailsScreen(id, title, category, imageUrls,
+              trailerVideoUrl, wholeItem, token);
         },
       ),
     );
@@ -44,10 +50,21 @@ class ItemWidget extends StatelessWidget {
   Widget rateStars(double starCount) {
     List<Widget> starList = [];
     var count = starCount.round();
+
     for (var i = 0; i < count; i++) {
       starList.add(
         Icon(
           Icons.star_rate_rounded,
+          color: Colors.grey[400],
+          size: 20,
+        ),
+      );
+    }
+    if (count.toDouble() > starCount) {
+      starList.removeLast();
+      starList.add(
+        Icon(
+          Icons.star_half_rounded,
           color: Colors.grey[400],
           size: 20,
         ),
@@ -63,33 +80,9 @@ class ItemWidget extends StatelessWidget {
     );
   }
 
-  Future<void> addItemToFav(String token) async {
-    var url = Uri.parse(
-      "https://sl-cinema.herokuapp.com/user/cinema/wish-list/add?id=" + wholeItem.id,
-    );
-
-    try {
-      var response = await http.get(
-        url,
-        headers: {
-          HttpHeaders.authorizationHeader: "Bearer " + token,
-          "content-type": "application/json",
-        },
-      );
-
-      print(response.statusCode);
-      print(response.body);
-    } catch (err) {
-      print("error");
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
-    bool isFavourite = Provider.of<Items>(context, listen: true).getFavItems.contains(wholeItem);
-
     final authData = Provider.of<Auth>(context);
-    final isUserAuth = authData.isAuth;
 
     final token = authData.getToken.toString();
 
@@ -98,6 +91,7 @@ class ItemWidget extends StatelessWidget {
 
     return Container(
       child: Card(
+        color: Colors.black,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(15),
         ),
@@ -115,7 +109,8 @@ class ItemWidget extends StatelessWidget {
                         begin: Alignment.center,
                         end: Alignment.bottomCenter,
                         colors: [gradientStart, gradientEnd],
-                      ).createShader(bounds);
+                      ).createShader(Rect.fromLTRB(
+                          0, 20, bounds.width, bounds.height - 40));
                     },
                     blendMode: BlendMode.darken,
                     child: Container(
@@ -126,9 +121,8 @@ class ItemWidget extends StatelessWidget {
                           image: NetworkImage(imageUrls),
                           fit: BoxFit.cover,
                         ),
-                        borderRadius: BorderRadius.only(
-                          topLeft: Radius.circular(15),
-                          topRight: Radius.circular(15),
+                        borderRadius: BorderRadius.all(
+                          Radius.circular(15),
                         ),
                       ),
                       // child: ClipRRect(
@@ -147,17 +141,18 @@ class ItemWidget extends StatelessWidget {
                     ),
                   ),
                   Positioned(
-                    bottom: 4,
+                    bottom: 20,
                     left: 3,
                     // right: 20,
                     child: Container(
                       width: 250,
-                      padding: EdgeInsets.symmetric(vertical: 5, horizontal: 25),
+                      padding:
+                          EdgeInsets.symmetric(vertical: 5, horizontal: 25),
                       child: Text(
                         title,
                         style: TextStyle(
                           fontWeight: FontWeight.bold,
-                          fontSize: 24,
+                          fontSize: 27,
                           color: Colors.grey[400],
                         ),
                         softWrap: true,
@@ -166,18 +161,42 @@ class ItemWidget extends StatelessWidget {
                     ),
                   ),
                   Positioned(
-                    bottom: 4,
+                    bottom: 0,
+                    left: 3,
+                    // right: 20,
+                    child: Container(
+                      width: 250,
+                      padding:
+                          EdgeInsets.symmetric(vertical: 5, horizontal: 25),
+                      child: Text(
+                        wholeItem.genres[0],
+                        style: TextStyle(
+                          fontWeight: FontWeight.w600,
+                          fontSize: 15,
+                          color: Colors.grey[400],
+                        ),
+                        softWrap: true,
+                        overflow: TextOverflow.fade,
+                      ),
+                    ),
+                  ),
+                  Positioned(
+                    bottom: 20,
                     left: 230,
                     // right: 20,
                     child: Container(
                       width: 300,
-                      padding: EdgeInsets.symmetric(vertical: 5, horizontal: 25),
+                      padding:
+                          EdgeInsets.symmetric(vertical: 5, horizontal: 25),
                       child: rateStars(wholeItem.ratings),
                     ),
                   ),
                 ],
               ),
             ),
+            SizedBox(
+              height: 20,
+            )
             // Container(
             //   child: Column(
             //     children: <Widget>[
@@ -206,91 +225,6 @@ class ItemWidget extends StatelessWidget {
             //     Text("Comment Count"),
             //   ],
             // ),
-            Padding(
-              padding: EdgeInsets.all(MediaQuery.of(context).size.height * 0.01),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                children: <Widget>[
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      IconButton(
-                        onPressed: null,
-                        icon: Icon(Icons.thumb_up),
-                      ),
-                      Text("like"),
-                    ],
-                  ),
-                  InkWell(
-                    onTap: () {
-                      !isUserAuth
-                          ? Navigator.of(context).push(
-                              MaterialPageRoute(
-                                builder: (_) {
-                                  return SignButtonScreen();
-                                },
-                              ),
-                            )
-                          : Navigator.of(context).push(
-                              MaterialPageRoute(builder: (_) {
-                                return CommentScreen(id, imageUrls, wholeItem, token);
-                              }),
-                            );
-                    },
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: <Widget>[
-                        IconButton(
-                          onPressed: () {
-                            !isUserAuth
-                                ? Navigator.of(context).push(
-                                    MaterialPageRoute(
-                                      builder: (_) {
-                                        return SignButtonScreen();
-                                      },
-                                    ),
-                                  )
-                                : Navigator.of(context).push(
-                                    MaterialPageRoute(
-                                      builder: (_) {
-                                        return CommentScreen(id, imageUrls, wholeItem, token);
-                                      },
-                                    ),
-                                  );
-                          },
-                          icon: Icon(
-                            Icons.comment,
-                            color: Colors.grey,
-                          ),
-                        ),
-                        Text("Review  "),
-                      ],
-                    ),
-                  ),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      IconButton(
-                        onPressed: () => {
-                          // favBtnTap(isFavourite, context, wholeItem),
-                          addItemToFav(token),
-                        },
-                        icon: isFavourite
-                            ? Icon(
-                                Icons.favorite,
-                                color: Colors.red,
-                              )
-                            : Icon(
-                                Icons.favorite,
-                                color: Colors.grey,
-                              ),
-                      ),
-                      Text("Favourite"),
-                    ],
-                  )
-                ],
-              ),
-            ),
           ],
         ),
       ),
